bot/requirements.txt
python-telegram-bot==12.1.1
requests==2.22.0